🎨 Abstract client component to separate component
